# src/main/resources/openapi/batchPrint.yaml
openapi: 3.0.0
info:
  title: Batch printing
  version: v1
paths:
  /print/entries:
    parameters:
      - $ref: "#/components/parameters/okapi_tenant"
      - $ref: "#/components/parameters/okapi_token"
      - $ref: "#/components/parameters/okapi_url"
      - $ref: "#/components/parameters/okapi_user"
    get:
      description: >
        Get batch printing entries with optional CQL query.
        X-Okapi-Permissions must include batch-print.entries.collection.get
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/query"
      operationId: getPrintEntries
      responses:
        "200":
          description: Print entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/entries"
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    delete:
      description: >
        Delete batch printing entries by  comma separated IDs.
        X-Okapi-Permissions must include batch-print.entries.collection.delete
      parameters:
        - $ref: "#/components/parameters/ids"
      operationId: deletePrintEntries
      responses:
        "204":
          description: Print entries deleted
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
    post:
      description: >
        Create print entry.
        X-Okapi-Permissions must include batch-print.entries.item.post
      operationId: postPrintEntry
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/entry"
      responses:
        "204":
          description: Print entry created
        "400":
          $ref: "#/components/responses/trait_400"
        "403":
          $ref: "#/components/responses/trait_403"
        "413":
          $ref: "#/components/responses/trait_413"
        "500":
          $ref: "#/components/responses/trait_500"
  /print/entries/{id}:
    parameters:
      - $ref: "#/components/parameters/okapi_tenant"
      - $ref: "#/components/parameters/okapi_token"
      - $ref: "#/components/parameters/okapi_url"
      - $ref: "#/components/parameters/okapi_user"
      - in: path
        name: id
        required: true
        description: Print entry identifier
        schema:
          type: string
          format: uuid
    get:
      description: >
        Get print entry by id.
        X-Okapi-Permissions must include batch-print.entries.item.get
      operationId: getPrintEntry
      responses:
        "200":
          description: Print entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/entry"
        "400":
          $ref: "#/components/responses/trait_400"
        "403":
          $ref: "#/components/responses/trait_403"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    delete:
      description: >
        Delete print entry.
        X-Okapi-Permissions must include batch-print.entries.item.delete
      operationId: deletePrintEntry
      responses:
        "204":
          description: Print entry deleted
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    put:
      description: >
        Update print entry.
        X-Okapi-Permissions must include batch-print.entries.item.put
      operationId: updatePrintEntry
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/entry"
      responses:
        "204":
          description: Print entry update
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /mail:
    parameters:
      - $ref: "#/components/parameters/okapi_tenant"
      - $ref: "#/components/parameters/okapi_token"
      - $ref: "#/components/parameters/okapi_url"
      - $ref: "#/components/parameters/okapi_user"
    post:
      description: >
        Send mail to create print entry.
        X-Okapi-Permissions must include batch-print.entries.mail.post
      operationId: saveMail
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/messageRequest"
      responses:
        "200":
          description: Print entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/messageResponse"
        "400":
          $ref: "#/components/responses/trait_400"
        "403":
          $ref: "#/components/responses/trait_403"
        "500":
          $ref: "#/components/responses/trait_500"
  /print/batch-creation:
    parameters:
      - $ref: "#/components/parameters/okapi_tenant"
      - $ref: "#/components/parameters/okapi_token"
      - $ref: "#/components/parameters/okapi_url"
      - $ref: "#/components/parameters/okapi_user"
    post:
      description: >
        Send mail to create print entry.
        X-Okapi-Permissions must include batch-print.print.write
      operationId: createBatch
      responses:
        "204":
          description: Print entry created
        "400":
          $ref: "#/components/responses/trait_400"
        "403":
          $ref: "#/components/responses/trait_403"
        "500":
          $ref: "#/components/responses/trait_500"
components:
  parameters:
    okapi_tenant:
      in: header
      name: X-Okapi-Tenant
      description: Okapi Tenant
      required: true
      schema:
        type: string
    okapi_token:
      in: header
      name: X-Okapi-Token
      description: Okapi Token
      required: false
      schema:
        type: string
    okapi_url:
      in: header
      name: X-Okapi-Url
      description: Okapi URL
      required: false
      schema:
        type: string
    okapi_user:
      in: header
      name: X-Okapi-User-Id
      description: Okapi user identifier
      required: false
      schema:
        type: string
    query:
      in: query
      name: query
      description: CQL query
      required: false
      schema:
        type: string
    offset:
      in: query
      name: offset
      description: Skip over number of elements (default is first element)
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    limit:
      in: query
      name: limit
      description: Limit the number of elements returned in the response
      required: false
      schema:
        type: integer
        default: 10
        minimum: 0
    ids:
      in: query
      name: ids
      description: Comma-separated IDs of items
      required: true
      schema:
        type: string
  responses:
    trait_400:
      description: Bad request
      content:
        text/plain:
          schema:
            type: string
            example: Invalid JSON in request
        application/json:
          schema:
            type: object
            example: {"error":"Invalid JSON in request"}
    trait_403:
      description: Forbidden
      content:
        text/plain:
          schema:
            type: string
            example: Insufficient permissions to access resource
        application/json:
          schema:
            type: object
            example: {"error":"Insufficient permissions to access resource"}
    trait_404:
      description: Not Found
      content:
        text/plain:
          schema:
            type: string
            example: Identifier 596d9f60-cda3-44d2-a4a1-2f48b7d4d23c not found
        application/json:
          schema:
            type: object
            example: {"error":"Identifier 596d9f60-cda3-44d2-a4a1-2f48b7d4d23c not found"}
    trait_413:
      description: Payload Too Large
      content:
        text/plain:
          schema:
            type: string
    trait_500:
      description: Internal error
      content:
        text/plain:
          schema:
            type: string
            example: Internal server error, contact administrator
  schemas:
    entries:
      description: Print entries response,
      type: object
      properties:
        items:
          description: List of print entries
          type: array
          items:
            $ref: "#/components/schemas/entry"
        resultInfo:
          $ref: "#/components/schemas/resultInfo"
      additionalProperties: false
      required:
        - items
    entry:
      description: Print entry
      type: object
      properties:
        id:
          description: Identifier
          type: string
          format: uuid
        created:
          type: string
          description: Creation date time
        type:
          type: string
          description: Print entry type
        sortingField:
          type: string
          description: Print entry sorting field
        content:
          type: string
          description: Print entry content
      additionalProperties: false
      required:
        - id
        - created
        - type
        - content
    resultInfo:
      description: Common result set information for streaming response
      type: object
      properties:
        totalRecords:
          description: Total number of entries in response
          type: integer
        diagnostics:
          description: Diagnostics for response
          type: array
          items:
            type: object
            properties:
              message:
                description: single diagnostic message
                type: string
        facets:
          type: array
          description: Array of facets
          items:
            type: object
            description: A facet
            properties:
              facetValues:
                type: array
                description: Array of facet values
                items:
                  type: object
                  description: A facet value
                  properties:
                    count:
                      type: integer
                      description: Count of facet values
                    value:
                      description: Value Object
              type:
                type: string
                description: Type of facet
      additionalProperties: false
    messageRequest:
      description: The mail message that can be sent via mail
      title: Mail Entity Schema
      type: object
      properties:
        notificationId:
          description: notification ID
          type: string
        from:
          description: sender's address
          type: string
        to:
          description: receiver's address
          type: string
        outputFormat:
          description: "format type: `text/html` or `text/plain`"
          type: string
        body:
          description: text of email
          type: string
      additionalProperties: true
      required:
        - body
    messageResponse:
      description: Mail message response
      type: object
      properties:
        id:
          description: Identifier
          type: string
          format: uuid
      additionalProperties: true
      required:
        - id